refactor(components): log button으로 전환
